got one post button to hide when modal displayed.... now for that pesky little widget component........
