+++ src/Components/Widget.css
@@ -1,10 +1,11 @@
 .widgets {
-	flex: 0.7 1;
+	flex: 0.7;
 	height: 100vh;
 	position: sticky;
 	top: 0%;
 	overflow: auto;
-  width: 500px;
+	width: 500px;
+	z-index: 0;
 }
 
 .widgets__input {
