+++ src/Screens/Home.jsx
@@ -19,8 +19,8 @@ function Home() {
 
 	return (
 		// BEM
-    <div className="home">
-      <Sidebar />
+		<div className="home">
+			<Sidebar />
 			<Feed />
 			<Widget />
 		</div>
